feat: added support for all files type
This change added two new settings. The first one is a list of MIME types that are resizable (on which the w and h query parameter will apply) and the second one is the list of allowed MIME file types to be uploaded. If an user tries to upload a file, and the MIME type of this file isn't in the allowed MIME types list, an error will be raised